+++ include/IR/IRBuilder.h
@@ -3,22 +3,37 @@
 #include <napi.h>
 #include <llvm/IR/IRBuilder.h>
 #include "IR/Value.h"
+#include "IR/ConstantInt.h"
+#include "IR/Type.h"
+#include "IR/IntegerType.h"
 
 #define binOpFactoryMacro(binOpFuncType, extraArgs...) \
 template<binOpFuncType method> \
 Napi::Value binOpFactory(const Napi::CallbackInfo &info) { \
     Napi::Env env = info.Env(); \
-    size_t argsLen = info.Length(); \
-    bool flag = argsLen >= 2 && Value::IsClassOf(info[0]) && Value::IsClassOf(info[1]); \
-    if (argsLen >= 3 && flag && info[2].IsString() || argsLen == 2 && flag) { \
+    size_t argsLen = info.Length();                    \
+    if(argsLen >= 2 && Value::IsClassOf(info[0]) && Value::IsClassOf(info[1])) { \
         llvm::Value *lhs = Value::Extract(info[0]); \
         llvm::Value *rhs = Value::Extract(info[1]); \
-        const std::string &name = argsLen >= 3 ? info[2].As<Napi::String>().Utf8Value() : ""; \
-        llvm::Value *result = (builder->*method)( lhs, rhs, name, ##extraArgs); \
-        return Value::New(env, result); \
+        if(argsLen == 2 || argsLen >= 3 && info[2].IsString()) { \
+            const std::string &name = argsLen >= 3 ? info[2].As<Napi::String>().Utf8Value() : ""; \
+            llvm::Value *result = (builder->*method)( lhs, rhs, name, ##extraArgs); \
+            return Value::New(env, result); \
+        } \
     } \
     throw Napi::TypeError::New(env, "IRBuilder createBinaryOperation needs to be called with: (lhs: Value, rhs: Value, name?: string)"); \
-} \
+}
+
+#define getIntFactoryMacro(funcType) \
+template<funcType method> \
+Napi::Value getIntFactory(const Napi::CallbackInfo &info) { \
+    Napi::Env env = info.Env(); \
+    if (info.Length() == 0 || !info[0].IsNumber()) { \
+        throw Napi::TypeError::New(env, "IRBuilder.#funcType need to be called with (value: number)"); \
+    } \
+    unsigned value = info[0].As<Napi::Number>(); \
+    return ConstantInt::New(env, (builder->*method)(value)); \
+}
 
 typedef llvm::IRBuilder<> LLVMIRBuilder;
 
@@ -30,6 +45,19 @@ typedef llvm::Value *(llvm::IRBuilderBase::*BinaryOperation)(llvm::Value *, llvm
 
 typedef llvm::Value *(llvm::IRBuilderBase::*BinaryOperationWithBool)(llvm::Value *, llvm::Value *, const llvm::Twine &, bool isExact);
 
+typedef llvm::ConstantInt *(llvm::IRBuilderBase::*GetBoolean)();
+
+typedef llvm::ConstantInt *(llvm::IRBuilderBase::*GetInt8)(uint8_t);
+
+typedef llvm::ConstantInt *(llvm::IRBuilderBase::*GetInt16)(uint16_t);
+
+typedef llvm::ConstantInt *(llvm::IRBuilderBase::*GetInt32)(uint32_t);
+
+typedef llvm::ConstantInt *(llvm::IRBuilderBase::*GetInt64)(uint64_t);
+
+typedef llvm::IntegerType *(llvm::IRBuilderBase::*GetIntType)();
+
+typedef llvm::Type *(llvm::IRBuilderBase::*GetType)();
 
 class IRBuilder : public Napi::ObjectWrap<IRBuilder> {
 public:
@@ -73,4 +101,40 @@ private:
     binOpFactoryMacro(BinaryOperation)
 
     binOpFactoryMacro(BinaryOperationWithBool, false)
+
+    Napi::Value getInt1(const Napi::CallbackInfo &info);
+
+
+    template<GetBoolean method>
+    Napi::Value getBoolFactory(const Napi::CallbackInfo &info) {
+        return ConstantInt::New(info.Env(), (builder->*method)());
+    }
+
+    getIntFactoryMacro(GetInt8)
+
+    getIntFactoryMacro(GetInt16)
+
+    getIntFactoryMacro(GetInt32)
+
+    getIntFactoryMacro(GetInt64)
+
+    Napi::Value getIntN(const Napi::CallbackInfo &info);
+
+    Napi::Value getInt(const Napi::CallbackInfo &info);
+
+    template<GetIntType method>
+    Napi::Value getIntTypeFactory(const Napi::CallbackInfo &info) {
+        return IntegerType::New(info.Env(), (builder->*method)());
+    }
+
+    template<GetType method>
+    Napi::Value getTypeFactory(const Napi::CallbackInfo &info) {
+        return Type::New(info.Env(), (builder->*method)());
+    }
+
+    Napi::Value getIntNTy(const Napi::CallbackInfo &info);
+
+    Napi::Value getInt8PtrTy(const Napi::CallbackInfo &info);
+
+    Napi::Value getIntPtrTy(const Napi::CallbackInfo &info);
 };