+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.2 (2021-03-26)
+
+### Features
+
+- add Class Argument
+- add ConstantFP.getNaN
+- add Function.arg_size and Function.getArg
+- delete Class FunctionCallee
+- supplement more APIs of Class IRBuilder
+- add more test cases
+- add type signature for new APIs
+
+
 ## 0.0.1 (2021-03-26)
 
 This is the first release of `llvm-bindings`.